Add files via upload
Here is a picture of the topology. It is still very much so a work in progress and I plan to add detailed updates and instructions every once in a while. Currently nothing apart from hostnames has been configured. The current plan is to have VLANs, Trunking, RPVST+, Port Channel, NAT, ACL and etc. configured on this network. Not only that but this is a way for me to get used to the idea of three tier networks. I plan to make similar networks in GNS3 in the future.